Change server and client component ordering
